+++ settings.gradle
@@ -6,3 +6,5 @@ include 'actuator'
 include 'consensus'
 include 'common'
 include 'example:actuator-example'
+include 'crypto'
+
